1. Load image .hex without debugger. 2. Assure integrity and authenticity of .hex when reprogramming 3. Assure integrity and authenticity of .hex when booting application 4. SHE lib: encryption (AES, ECB, CBC algorithm + code), CMAC, TRNG, PRNG, SecureBoot, KeyMng
**---------Challenge----------------**
Key Management System
Secure Key Storage

+++ SFBL_Tuan/DEMO1/Src/cmac.c
@@ -17,7 +17,7 @@ void Delay1(int num)
 }
 
 unsigned char MAC_Bootloader[16] = {0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00};
-bool Verify_MAC(unsigned char* K, unsigned char* M, uint32_t len, unsigned char* Reference_MAC)
+bool Verify_MAC(unsigned char* K, unsigned char* M, uint8_t len, unsigned char* Reference_MAC)
 {
 	AES_CMAC(K, M, len, MAC_Bootloader);
 	if (0 == memcmp(MAC_Bootloader, Reference_MAC, 16))
